fetch-agents/requirements.txt
uagents==0.12.0
requests==2.31.0
google-generativeai==0.3.2
python-dotenv==1.0.0
fastapi==0.109.0
uvicorn==0.27.0
